openaddresses: handle geojson file (#69)
CSV appears to have been deprecated for a while in OpenAddresses and has
recently been removed from global exports.

Note that I put a quite naive parser which expects to find a simple
point shape, we could use more advanced shapes and geos if the files get
more complicated in the future (eg. with polygons).